+++ extensions/vercast/src/pages/with-valid-token.tsx
@@ -2,15 +2,12 @@ import type { FC, PropsWithChildren } from "react";
 import isValidToken from "../utils/is-valid-token";
 import InvalidTokenView from "./details/invalid-token-view";
 import { usePromise } from "@raycast/utils";
-import { Detail, showToast, Toast } from "@raycast/api";
+import { Detail } from "@raycast/api";
 
 const WithValidToken: FC<PropsWithChildren<object>> = ({ children }) => {
   const { isLoading, error } = usePromise(isValidToken, [], {
-    onError() {
-      showToast({
-        style: Toast.Style.Failure,
-        title: "Invalid token detected. Please set one in the settings.",
-      });
+    failureToastOptions: {
+      title: "Invalid token detected. Please set one in the settings.",
     },
   });